+++ website/docs/data-display/bar-horizontal/examples/legend-and-pattern-fill.tsx
@@ -72,6 +72,7 @@ const Demo = () => {
         onMouseEnterItem={handleMouseEnter}
         onMouseLeaveItem={handleMouseLeave}
         patterns
+        aria-label={'Legend for the horizontal bar chart'}
       />
       <Plot
         data={data}