+++ semcore/mini-chart/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [0.20.1] - 2024-10-08
+
+### Fixed
+
+- Types for enhances.
+
 ## [0.20.0] - 2024-09-20
 
 ### Changed